biasgen buttons changed to transparent background so that they show their pressed state.
slider in AEViewer.form greatly improved so that it shows a live paused view of the underlying data.
resize button fixed so that the jpanel structure puts it in the corner and so that the slider fills the rest of the bottom row. buttons like biass, filters, etc made smaller as well.

RectangularClusterTracker.java: new field added to cluster for birth location to enable determination of total distance traveled since birth.

Goalie: new state added SLEEPING from which it is harder to acquire a ball for blocking, which should help with annoying distractions from background.

many log messages commented out throughout jAER.jar