fixes to reset timestamps on stereo pair more consistently. now the buffers are not flushed anymore and only a hardware reset is called for by vendor request. the data coming after the reset may still contain old (pre-reset) timetamps. Tmpdiff128StereoPair.java no longer applies heuristic to reset timestamps if they differ too much by acquisition time, since this hack didn't seem to work consistently in any case.
FilterFrame now uses the chip preferences node rather than a mixture of it's own preferences and the chip prefernces. It now also purges the list of preferred filters of the filters that cannot be constructed for one reason or another - thus classes which no longer exist are purged from the list.

BinocularRenderer now uses the instantaneous count of events in the packet to render, because the count obtained before the loop was changing, presumably because some other thread was touching the data (cause not determined).