irotate filter now includes an arbitrary anglebut this doesn't work yet.
Chip2D now has a preferred display method that is restored on startup. This implementation takes account of the fact that not all display methods have been constructed for a ChipCanvas at the time the prefeered one is called for, and instantiates a new DisplayMethod instance to fill in the required instance. The list is later chacked to make sure the display method is not duplicated.

CochleaGenderClassifier finds the sex but not accurately yet.