added mass to toString for RCT cluster, for debugging slot car tracking
SlotcarTrack now does a *signed* distance computation, and it happens that more negative is on the "inside* of the track model for CCW tracks when viewed from above with increasing segment numbers.

minor changes to evolutionaryThrottleController to added an "editThrottleChange" for contolling mouse editing of throttle amount.
improvements in tooltips for slotcar classes.