fixed bug in voting for car cluster from all cluster where voting index was not being incremented.
implemented and used the new tracker-based track extraction algorithm to extract a very accurate track pasadena2-tracker.track in the telluride slot10 svn.

implemented various options to control display of track and trackerPositions stored in the TrackdefineFilter.

changed how the track is updated by addPoint - now you need to manually call updateTrack() on the track to avoid storm of callbacks.