committed changed from pedro zucarello to fix biasgen.c for ledit 14.1, according to the following comments:
Other things: I have been working with the bias generator macros for the
Tanner tools, and I found some bugs at macro for the fixed biases. I
corrected the bugs and now the script works fine under 14.1 version of
LEdit. Here I send you the file with the corrected version of the macro
in case someone wants to use it. All corrections are labeled as PDZ in
case you want to take a look at them.