C++ interface to the DVS128
Compiled with QtCreator, Qt 4.8.1 and MSVC2010. Example program starts a widget which shows the raw camera output on screen.

Usage of the interface:
To get events, inherit from Evenprocessorbase and implement processEvent(Event e). Make an instance of DVS128interface, hand over a reference of your eventprocessor class and call startReading(). If you want to send a vendor requests to the camera, like for changing the biases, use sendVendorRequest in dvs128interface.