Fremes are the event based equivalents to frames that get updated on the arrival of an event. A freme can contain any type of object. The objects in the freme class are stored in an array. To access the objects the according methods of the freme class should be used.
A freme extractor extracts and maintains a freme from an event stream. I is also capable of displaying its freme. Fremes can be processed hierarchically similar to filters with one freme extractor getting the freme from a previous stage.

A first illustration of how to implement a freme as well as a freme extractor is the orientation freme extractor that stores the latest orientations of a pixel in a freme. The orientation is calculated with a cross kernel on the latest timestamps. Along an edge, events are assumed to have similar timestamp whereas accross the have big differences.