Started some code cleanup. Mainly reformating comments, realigning definitions and introduced code-folding for the most obvious getter/setter methods.
This code is seriously cluttered!
I think I could understand most of it but because of the sheer amount of missing structure in the code its really hard!
Could use some more cleanup!
Especially all that probabliy old commented code should be deleted if no longer needed. But Im not Adept enough to decide weather its needed at some point or not