Major code cleanup. 	Grouping parameters, code folding, getter/setter consistency
New features:
	-Added 'Targeting" of the PanTilt and jittering while targeting.
	 This means that one does not need to set the pantilt values directly but instead sets a target 
	 that will be moved towards in a linear fashion with speed proportional to distance.
	-Added listeners to pantilt values, target and jittertarget such that it can be evaluated by downstream filters.
	-Changed the GUI such that it shows the Target, JitterTarget and position of pantilt if requested
	-Pan/Tilt values can be set in the GUI and the Aimer directly via textinput

More will be added/changed in the future. I am not changing existing methods where possible so to make sure that other downstream classes are still usable but there is a difference weather a lasepointer is present or not and the whole calibration is different for the camera on the pantilt. A new calibration is currently developed