Major overhaul of the direction selective filter.
There was a bug with the 'type' that caused the horizontal and vertical motion direction to be much more prominent, as on AND off events where both safed in the same TimeStampMap instead of seperate ones.

Also a lot of code cleaning, commenting and code folding.
No new functionality other than the bugfix. Code should run smoother and faster, as some loops and checks where inefficient.