Introduced a boolean flag 'hasDirection' to 'MotionOrientationEvent' that flags whether a valid Motion Direction was found. Introduced the switch 'passAllEvents' in AbstractDirectionSelectiveFilter to allow for all Events to be passed if the user chooses so. This makes the Filter actually a Labeler.
Introduced functionality for passing all Events in DvsDirectionSelectiveFilter. When the switch is set to 'false' behaviour is the same as before.

Also introduced a check in DvsDirectionSelectiveFilter to check that incoming events have 'hasOrientation' flagged as true. This allows the Filter to work even if the flag 'passAllEvents' in the OrientationFilter is switched on.