changed FX2 I2C implementation of IMU integration so that IMU samples can now be turned off (and are turned off) when they are not shown. The firmware now sets an internal flag that prevents sending IMU samples to host when showIMU is disabled on SBRet10 chip class. Also removed injection of fake external pin events so that now if IMU samples are turned off, there are no special events, and if they are turned on, there is about 1 special event per ms of displayed packet.
Changed names of some constants in ApsDvsChip class to make more sense, and added conditionals (not yet fully tested) in ApsDvsHardwareInterface.AEReader.translateEvents and SBRet10.EventExtractor to test bits for IMU sample. This code is not yet working I think.

Also changed logic of computing whether to update statistics text in AEViewer so that now stats are updated if they have not been for the last 100ms, if a frame is rendered.