Adjusted several filters to use the new drawGL class. In DirectionSelective filter the velocities are now drawn as drawGLVectors
In ClusterTracker the cluster now paints itself completely independent of the ClusterTrackerClass. This means that it uses the drawGL.Box, drawGL.Ellipse and drawGL.Line
The depending classes have been adjusted to use the drawGL commands as well such that the drawEllpise and drawBox method of the CLustertracker was obsolete and subsequently deleted.

Also included the drawGL methods in several other classes